update :: moved Global
