+++ src/main/java/com/htmake/htbot/discord/commands/battle/cache/MonsterStatusCache.java
@@ -1,6 +1,6 @@
 package com.htmake.htbot.discord.commands.battle.cache;
 
-import com.htmake.htbot.cache.CacheManager;
+import com.htmake.htbot.global.cache.CacheManager;
 import com.htmake.htbot.discord.commands.battle.data.MonsterStatus;
 
 public class MonsterStatusCache {
