+++ src/main/java/com/htmake/htbot/discord/commands/player/PlayerCommand.java
@@ -2,7 +2,7 @@ package com.htmake.htbot.discord.commands.player;
 
 import com.htmake.htbot.discord.commands.player.event.PlayerJoinEvent;
 import com.htmake.htbot.domain.player.enums.Job;
-import com.htmake.htbot.unirest.HttpClient;
+import com.htmake.htbot.global.unirest.HttpClient;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
 import kotlin.Pair;
